py/packages/sdk/pyproject.toml
[project]
name = "trysoma_sdk"
version = "0.0.4"
description = "Python SDK for Soma"
requires-python = ">=3.10"
license = "MIT"
keywords = ["soma", "sdk", "agent", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "trysoma_sdk_core",
    "trysoma_api_client",
    "pydantic>=2.0",
    "watchfiles>=0.21.0",
    "httpx>=0.27.0",
    "restate-sdk>=0.12.0",
    "hypercorn>=0.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["trysoma_sdk"]

[project.urls]
Homepage = "https://github.com/trysoma/soma"
Repository = "https://github.com/trysoma/soma"


[tool.mypy]
python_version = "3.10"
strict = true
mypy_path = "../api_client"

[[tool.mypy.overrides]]
module = "trysoma_api_client.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "trysoma_sdk_core.*"
ignore_errors = true
ignore_missing_imports = true

[tool.pyright]
extraPaths = ["../api_client"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.uv.sources]
trysoma_sdk_core = { path = "../../../crates/sdk-py" }
trysoma_api_client = { workspace = true }
